+++ changelog.md
@@ -1,6 +1,7 @@
 # Incrementum changelog:
 
 - Changed the name to "Incrementum" and replaced all instances of "mod" with "game". game.js has been renamed incrementum.js, and mod.js is now game.js.
+- Added linear and quadratic prestige types, improved the prestige type system.
 - Upgrade effectDisplay and grid tooltip no longer display if they return "".